Cleaner petitions
Cleaned up the petitions quite a bit. Added the ability to modify the
petition form for future modularity. Converted the closedate times in
petition viewer to dhms() format, so you can see how long it has been
since an update. Removed formatting on the date the petitions were sent
in, and now the page info displays as json output.