Reorganized gamelog, fixed internal view of petitions.
The internal view of petitions will now show who last updated the
petition if it was actually updated. The game log will now show the name
before the message with a timestamp, as well as the number of items per
category on the side bar. In the future, there will be a show/hide when
clicking on those categories in the navigation.